pyproject.toml
[tool.poetry]
name = "geec"
version = "0.0.3"
description = "Program to calculate gravity and gravity gradients due to irregularly shaped bodies."
authors = ["Julien Paul <fjulien.paul@umontpellier.fr>"]
repository = "https://github.com/julienpaul/geec"
documentation = "https://julienpaul.github.io/geec/"
readme = "README.md"
packages = [
  {include = "geec"}
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
typer = "^0.9.0"
loguru = "^0.7.0"
numpy = ">=1.24.2"
scipy = ">=1.10.1"
pandas = "^2.0.1"
rich = "^13.3.5"
confuse = "^2.0.1"
plotly = "^5.15.0"
perfplot = "^0.10.2"
pymap3d = "^3.0.1"
pygeodesy = "^23.6.12"
pyglm = "^2.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
deptry = ">=0.6.4"
mypy = ">=0.981"
pre-commit = ">=2.20.0"
tox = ">=3.25.1"
black = ">=23.3.0"
ruff = ">=0.0.274"
line-profiler = "^4.0.3"
poetry2conda = "^0.3.0"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.3"
mkdocstrings = {extras = ["python"], version = "^0.22.0"}
mkdocs-material = "^9.1.18"
# mkdocstrings-python = "^1.1.2"
griffe = "0.21.0"
mkdocs-include-markdown-plugin = "^4.0.4"

[tool.poetry.scripts]
geec = "geec.cli:app"

[tool.poetry2conda]
name = "geec-env"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
preview = true

[tool.mypy]
files = ["geec"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.ruff]
target-version = "py311"
line-length = 88
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["geec"]
omit = ["tests/test_timeit.py"]

[[tool.poetry_bumpversion.replacements]]
files = ["geec/__init__.py"]
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.poetry_bumpversion.replacements]]
files = ["tests/test_geec.py"]
search = '__version__ == "{current_version}"'
replace = '__version__ == "{new_version}"'

[tool.deptry]
extend_exclude = [".*/test/"]
